+++ app/controllers/application_controller.rb
@@ -1,6 +1,4 @@
 class ApplicationController < ActionController::Base
-  include DeliverNowWithErrorHandling
-
   protect_from_forgery with: :exception
   before_action :authenticate_user!
   before_action :set_admin_status