+++ app/controllers/offers_controller.rb
@@ -12,6 +12,7 @@ class OffersController < PublicController
   def create
     submission = SubmissionForm.build submission_params
     if submission.save
+      EmailNewSubmission.run! submission: submission, user: current_user
       redirect_to contribution_thank_you_path, notice: 'Offer was successfully created.'
     else
       render_form(submission)